feat: 로그인 후 유저 정보 가져오기
- LogIn, Sign Up page/users로 이동
- TopBar에 있던 유저 정보 컨테이너 파일 분리
- 유저 정보 로컬스토리지에 저장
