+++ src/component/contexts/usersContext.jsx
@@ -9,12 +9,15 @@ export const UsersProvider = ({ children }) => {
 
   // 페이지 새로고침 해도 토큰 있으면 로그인 상태 유지
   useEffect(() => {
-      const token = localStorage.getItem('token');
+    const token = localStorage.getItem('token');
+    const user = localStorage.getItem('user');
     
       if (token) {
-          setIsUser(true);
+        setIsUser(true);
+        setUser(JSON.parse(user));
       } else {
-          setIsUser(false);
+        setIsUser(false);
+        setUser(null);
       }
   }, []);
